minor tweaks
In the home/index.html.erb the carousel or something along the carousel
border was blocking the mouse from highlighting and clicking on the menu
items.  moved it down slightly.  The News and Testimonial row, the
Testimonial was not quite aligned witht he right.  Needed a span between
5 and 6 and created 0 in the bootstrap.  If you have a better solution,
have at it. :-)